STL-4796 Change dlid size in struct ib_ah_attr from 16 to 32 bits
OPA devices can support larger number of nodes than 16 bit lids
can accommodate. This change increases the dlid size from 16 to
32 bits in the kernel so kernel clients such as ib_sa can set up
address handles with 32 with dlid.
UAPI is unchanged and is still only 16 bits.

+++ hfi2/verbs/rc.c
@@ -998,7 +998,7 @@ void hfi2_send_rc_ack(struct hfi2_qp *qp, int is_fecn)
 	 */
 	lrh0 |= (sc5 & 0xf) << 12 | (qp->remote_ah_attr.sl & 0xf) << 4;
 	hdr.lrh[0] = cpu_to_be16(lrh0);
-	hdr.lrh[1] = cpu_to_be16(qp->remote_ah_attr.dlid);
+	hdr.lrh[1] = cpu_to_be16((u16)qp->remote_ah_attr.dlid);
 	hdr.lrh[2] = cpu_to_be16(hwords + SIZE_OF_CRC);
 	hdr.lrh[3] = cpu_to_be16(ibp->ppd->lid |
 				  qp->remote_ah_attr.src_path_bits);
@@ -2146,7 +2146,7 @@ void hfi2_rc_rcv(struct hfi2_qp *qp, struct hfi2_ib_packet *packet)
 
 	if (unlikely(bth1 & (HFI1_BECN_SMASK | HFI1_FECN_SMASK))) {
 		if (bth1 & HFI1_BECN_SMASK) {
-			u16 rlid = qp->remote_ah_attr.dlid;
+			u32 rlid = qp->remote_ah_attr.dlid;
 			u32 lqpn, rqpn;
 
 			lqpn = qp->ibqp.qp_num;
